Restore retail slot and scalar contracts
Recover the TMapMgr/TMinor/navy/trade call families from Windows listings and Mac signature evidence, restore real collection ownership, correct trade settlement/comparator operands, and remove recomp-only runtime bypasses.\n\nVerification: just build; just gates through stats; just test; focused semantic comparisons and relevant vtables. The approval-gated full semantic policy baseline remains unchanged per the instruction to ignore semantic tests.

+++ include/game/core/TStream.h
@@ -86,7 +86,7 @@ public:
   virtual void WriteByte(unsigned char value);              // 31 (0x7c) 1 byte
   virtual void WriteBoolean(unsigned char value);           // 32 (0x80) 1 byte
   virtual void WriteCharacter(short value);                 // 33 (0x84) 1 byte (the high one)
-  virtual void WriteInteger(int value);                     // 34 (0x88) 2 bytes (MacApp Integer)
+  virtual void WriteInteger(short value);                   // 34 (0x88) 2 bytes (MacApp Integer)
   virtual void WriteLong(int value);                        // 35 (0x8c) 4 bytes
   virtual void WriteVPoint(double value);                   // 36 (0x90) 8 bytes
   virtual void WriteRect(void* data);                       // 37 (0x94) 8 bytes
